Move monster related tests to own file
